Code cleanup
- renamed apihandler package accordingly
- renamed ampq package to amqp10
- removed testing methods and loops from code
- renamed elasticsearch package accordingly
- moved elasticsearch unit test to tests package
- pinned elasticsearch container to 6.0.1 because we are
  incompatible with v7
