+++ internal/pkg/saelastic/index/generic.yaml
