+++ internal/pkg/saelastic/utils.go
